requirements.txt
asgiref==3.5.2
Django==4.1.3
django-froala-editor==4.0.16
django-tinymce==3.5.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.6
